+++ src/main/java/api/Client.java
@@ -1,3 +1,5 @@
+package api;
+
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.http.ContentType;
 import io.restassured.specification.RequestSpecification;
