+++ src/main/java/api/courierpackage/Courier.java
@@ -1,13 +1,18 @@
-public class CourierLogin {
+package api.courierpackage;
+
+public class Courier {
     private String login;
     private String password;
+    private String firstName;
 
-    public CourierLogin(String login, String password) {
+    public Courier(String login, String password, String firstName) {
         this.login = login;
         this.password = password;
+        this.firstName = firstName;
     }
 
-    public CourierLogin() {
+    public Courier() {
+
     }
 
     public String getLogin() {
@@ -24,5 +29,13 @@ public class CourierLogin {
 
     public void setPassword(String password) {
         this.password = password;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
     }
 }
