+++ src/main/java/api/courierpackage/CourierGenerator.java
@@ -1,3 +1,5 @@
+package api.courierpackage;
+
 public class CourierGenerator {
     public static Courier getDefault() {
         return new Courier("sdfsdf", "sdfsdf", "sdfsdf");
