+++ src/main/java/api/courierpackage/CourierLogin.java
@@ -1,16 +1,15 @@
-public class Courier {
+package api.courierpackage;
+
+public class CourierLogin {
     private String login;
     private String password;
-    private String firstName;
 
-    public Courier(String login, String password, String firstName) {
+    public CourierLogin(String login, String password) {
         this.login = login;
         this.password = password;
-        this.firstName = firstName;
     }
 
-    public Courier() {
-
+    public CourierLogin() {
     }
 
     public String getLogin() {
@@ -27,13 +26,5 @@ public class Courier {
 
     public void setPassword(String password) {
         this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
     }
 }
