+++ src/main/java/api/orderspackage/OrdersCancel.java
@@ -1,3 +1,5 @@
+package api.orderspackage;
+
 public class OrdersCancel {
     private int track;
