+++ src/main/java/api/orderspackage/OrdersClient.java
@@ -1,3 +1,6 @@
+package api.orderspackage;
+
+import api.Client;
 import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 
@@ -6,6 +9,7 @@ import static io.restassured.RestAssured.given;
 public class OrdersClient extends Client {
     private static final String PATH = "api/v1/orders";
     private static final String PATH_CANCEL = "api/v1/orders/cancel";
+
     @Step("/api/v1/orders - create")
     public ValidatableResponse create(Orders order) {
         return given()
@@ -15,6 +19,7 @@ public class OrdersClient extends Client {
                 .post(PATH)
                 .then();
     }
+
     @Step("/api/v1/orders/cancel")
     public ValidatableResponse delete(OrdersCancel ordersCancel) {
         return given()
@@ -24,7 +29,8 @@ public class OrdersClient extends Client {
                 .put(PATH_CANCEL)
                 .then();
     }
-    @Step("/api/v1/orders - get all Orders")
+
+    @Step("/api/v1/orders - get all OrdersPackage.Orders")
     public ValidatableResponse getOrders(Orders order) {
         return given()
                 .spec(getSpec())
